+++ lib/command/executor/pusher.rb
@@ -11,11 +11,25 @@ module PodPrebuild
 
     def run
       Pod::UI.step("Pushing cache") do
-        commit_message = "Update prebuilt cache".shellescape
-        git("add .")
-        git("commit -m '#{commit_message}'")
-        git("push origin #{@cache_branch}")
+        if @config.local_cache?
+          print_message_for_local_cache
+        else
+          commit_and_push_cache
+        end
       end
+    end
+
+    private
+
+    def print_message_for_local_cache
+      Pod::UI.puts "Skip pushing cache as you're using local cache".yellow
+    end
+
+    def commit_and_push_cache
+      commit_message = "Update prebuilt cache"
+      git("add .")
+      git("commit -m '#{commit_message}'")
+      git("push origin #{@cache_branch}")
     end
   end
 end